requirements.txt
asgiref==3.4.1
certifi==2022.6.15
charset-normalizer==2.0.12
Django==3.2.14
django-bootstrap-icons==0.8.1
django-ckeditor==6.3.2
django-js-asset==2.0.0
django-multiselectfield==0.1.12
gunicorn==20.1.0
idna==3.3
Pillow==8.4.0
# psycopg2==2.7.7
psycopg2-binary==2.9.3
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.10
whitenoise==5.3.0
